+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 The following is a curated list of changes in the Enact sandstone module, newest changes on the top.
 
+## [unreleased]
+
+### Fixed
+
+- `sandstone/Scroller` to handle focus moving properly when `editable` is given
+- `sandstone/Scroller` to not select disabled item in scroller when `editable` is given
+
 ## [2.7.4] - 2023-07-19
 
 ### Fixed