+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The following is a curated list of changes in the Enact sandstone module, newest
 
 - `sandstone/Scroller` prop `editable.initialSelected` to allow start edit mode with selected item
 
+### Changed
+
+- `sandstone/Scroller` back key behavior to match the latest UX when `editable` is given
+
 ### Fixed
 
 - `sandstone/Scroller` to handle focus moving properly when `editable` is given